--- src/rust/lqos_node_manager/src/tracker/cache/lqosd_stats.rs
@@ -1,5 +1,0 @@
-use once_cell::sync::Lazy;
-use std::sync::RwLock;
-
-pub static HOST_COUNTS: Lazy<RwLock<(u32, u32)>> =
-  Lazy::new(|| RwLock::new((0, 0)));
